allow overriding the DPI
RELNO allow pixel-perfect display of fonts on display-scaled displays. Until now these would appear blurry because Chrome makes px CSS units not really pixels, but an arbitrary scale of pixels based on the "looks like" specification in the display settings. To get pixel-perfect display, either at start-up or in a macro, set the dpi-fudge CSS variable to a fraction calculated with A / B, where A is the "looks like width" and B is the "native width" according to display settings. All px values will hence be multiplied by this factor before applied to a CSS property, and all Javascript-reported pixel sizes will be divided by this factor before processing.