split up hterm source files
There are an engine module and a client module. The separation is rough
for now but will be more rigid as time goes on.

The engine will eventually have syntax-neutral code only for managing
the terminal state, so it can be used as C code as well.

The client module is everything else and is true Javascript.