Refactor Screen class into the tm style
This was automated somewhat, though the diff was reviewed line-by-line.

This also includes the sed script which does the full refactor. There
are no manual modifications. Future classes will be easier, though the
next version of the script may have to process nested parenthesis in a
way that sed and regex's don't usually do.

This is because foo.bar has changed to bar(foo) in many contexts, and
this will continue. Eventually this will result in bar(foo) passed to
baz(...), and then baz having to be transformed as a single unit later.