RELNO automatically-assigned terminal IDs are now shorter and nearly guaranteed to be unique, avoiding the issue where starting a terminal with "laH T " will open a previously-started one. The automatic parent-child relationship of spawned terminals is no longer used.
RELNO allow logging for the dtach component instead of sending everything to /dev/null. This has to be turned on manually.

RELNO removed the "scratch" section of the /attach page. Instead, the "Common" section is now guaranteed to start a new terminal of that profile.

RELNO a "basic" profile is always available, but can also be configured just like regular profiles. This
corresponds to the defaultish but persistent behavior of the now-removed "scratch" feature.

RELNO ephemeral sessions now belong to the "basic" profile, so they can be configured.