Be hands-off on environment modification
There is no reason to sanitize the environment, so just stop. All we
need is to add QUERY_STRING to the parent's environment. Now the "run"
script does not need to search for which environment variables to
preserve.

Implement showenv, a debugging tool, in C. All CGI scripts will
eventually be removed and the logic moved to C.