index.html: use \n to delimit escapes
Would be nice to simplify websocket protocol implementation. To avoid
requiring the server to include the entire escape in a websocket frame,
make index.html tolerate escapes that span websocket messages.

To accomplish this, websocketd stops trimming the \n from frames. The
session binary continues to send \n, and index.html uses this to figure
when the frame is done. If the frame is not done, no escapes, neither 2-
character ones nor payload ones (\@...), will be processed.

RELNO will require restarting spawner manually

BUG=https://github.com/google/werm/issues/1