session.c: deframe inbound websock data
Make session.c do some websocket work rather than websocketd.

Reduces size of code base and gets closer to single-language.
See https://github.com/google/werm/issues/1

Show dtach termination status in the terminal properly escaped, as
needed. Error details for attach, in case of an error, are now written
to attach process' stderr rather than stdout, which causes it to appear
in websocketd's log, and which allows more idiomatic error handling
(perror).

dtach termination is shown colorized so it is easy to see. Original
dtach used the end-of-screen ANSI escape, which was not working for
Werm, so we no longer output it.