Remove unused websocketd and gorillaws code
This removes some undocumented (in Werm's README) flags in websocketd,
including those that check for same-origin. This is redundant with the
logic to check against cross-site fetching. If we want to do this
redundant check, we can add it back in the first-party implementation,
so add a TODO for that.

See https://github.com/google/werm/issues/1