RELNO basic Firefox support; remove unused code
Some things had to be loaded asynchronously to make Firefox work
properly. In the process of fixing, preference functionality that is
unused but was complicating the startup process has been deleted.

I don't know if inline images even work, but I removed the code showing
a UI which allowed the user to decide if they want them. This helped
simplify the preferences functionality and just replace it with a JS
object.