fix resize logic for DPI-enabled screens
This seems to work consistently across many DPI settings with various
per-page scaling settings. Meaning, there is no blur.

The key is to get the subpixel size of the window. Just reading
window.innerWidth or .innerHeight gives an integer-rounded value for the
terminal size in post-scaled pixels, so scaling that to physical pixels
will result in off-by-one values often. Use getComputedStyle to find the
non-rounded size.